+++ service/PostResultService.cs
@@ -5,17 +5,17 @@ namespace TelegrammService.service
     public class PostResultService
     {
 
-        public static PostResult getErrorPostResult(int apiId, string description)
+        public static PostResult getErrorPostResult(long apiId, string description)
         {
             return getPostResult(apiId, description, PostResultCode.ERROR);
         }
 
-        public static PostResult getOkPostResult(int apiId, string description)
+        public static PostResult getOkPostResult(long apiId, string description)
         {
             return getPostResult(apiId, description, PostResultCode.OK);
         }
 
-        public static PostResult getPostResult(int apiId, string description, PostResultCode postResultCode)
+        public static PostResult getPostResult(long apiId, string description, PostResultCode postResultCode)
         {
             PostResult postResult = new PostResult();
             postResult.apiId = apiId;
